# Conflicts:
#	Cargo.toml
#	src/modules/galaxy/models/planet_class.rs
#	src/modules/logs/content/log_event_content.rs
#	src/modules/logs/content/log_event_content/approach_settlement_event.rs
#	src/modules/logs/content/log_event_content/bounty_event.rs
#	src/modules/logs/content/log_event_content/cargo_depot_event.rs
#	src/modules/logs/content/log_event_content/community_goal_reward_event.rs
#	src/modules/logs/content/log_event_content/disembark_event.rs
#	src/modules/logs/content/log_event_content/dock_srv_event.rs
#	src/modules/logs/content/log_event_content/docking_requested_event.rs
#	src/modules/logs/content/log_event_content/embark_event.rs
#	src/modules/logs/content/log_event_content/fsd_target_event.rs
#	src/modules/logs/content/log_event_content/launch_srv_event.rs
#	src/modules/logs/content/log_event_content/liftoff_event.rs
#	src/modules/logs/content/log_event_content/mission_completed_event.rs
#	src/modules/logs/content/log_event_content/pay_bounties_event.rs

+++ src/modules/logs/content/log_event_content/liftoff_event.rs
@@ -21,7 +21,7 @@ pub struct LiftoffEvent {
 
     /// The id of the body the player lifting off from.
     #[serde(rename = "BodyID")]
-    pub body_id: u64,
+    pub body_id: u8,
 
     /// Whether the player is lifting off from a settlement.
     pub on_station: bool,
